Add .travis.yml for running "make distribution" tests.
OSX does not have fakeroot, so we avoid it.
This will mean that the archive files aren't owned by root.

Also when testing with clang Mk_dist didn't pass it onto the bootblocks
directory. It now does so; this means we can work around clang's warnings
that are sort of errors without actually fixing the code.